Captain's log
TypeScript decided to stop working in VS Code (but elsewhere) so
today I worked without it. I tried fixing it for half an hour, but then
went on to add the avatar upload.

Learned that Firebase doesn't let you set arbitrary user data. It does
let you set it in analytics! Which is pretty cool.

I'm wondering whether it's better to collect all this info ourselves (in
our own user model, in Rails...)

Yeah, that's what I think it is. Here's what's coming up next:

1. Fix this TypeScript-VS Code issue
2. Authenticate the current user on the Rails side
3. Create a route for checking the validity of a username
    3.1. Handle the entering-username use case
    3.2. Handle the submitting-but-someone-just-took-it use case
4. Set up a Points UI in the onboarding (maybe w/ Tooltips?). And playful animations!